requirements.txt
asgiref==3.5.0
Django==4.0.2
django-eav2==1.2.0
django-jsonfield-backport==1.0.4
django-tinymce==3.4.0
django-widget-tweaks==1.4.12
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-slugify==5.0.2
sqlparse==0.4.2
text-unidecode==1.3
